+++ EveUtils.Migrations.Server.Sqlite/Migrations/20260911124610_AddRunSignatureGroupSnapshot.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace EveUtils.Migrations.Server.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddRunSignatureGroupSnapshot : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "SignatureGroupSnapshot",
+                table: "Run",
+                type: "TEXT",
+                maxLength: 64,
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "SignatureGroupSnapshot",
+                table: "Run");
+        }
+    }
+}